Fix key error formatting
